Create component for hobbies
