+++ lib/interfaces/Hobby.tsx
@@ -1,0 +1,7 @@
+interface Hobby {
+  name: string;
+  description?: string; // Optional description
+  icon?: string; // Optional icon URL or class name
+}
+
+export default Hobby;
